requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
qrcode[pil]==7.4.2
Pillow==11.1.0
python-dotenv==1.0.1
pydantic[email]==2.10.6
pydantic-settings==2.7.0
aiofiles==23.2.1
